@@ -7,10 +7,12 @@
     "authors": [
         {
             "name": "Conn Warwicker"
+        }, {
+            "name": "Gwen Helleux"
         }
     ],
     "require": {
-        "php": ">=7.0.4"
+        "php": ">=7.2"
     },
     "require-dev": {
       "phpunit/phpunit": "dev-master"  
