@@ -4,111 +4,263 @@ namespace Waddle;
 
 class Lap
 {
-    
-    protected $totalTime; // Seconds
-    protected $totalDistance; // Metres
-    protected $maxSpeed; // Metres per second
-    protected $totalCalories;
-    protected $trackPoints = array();
-    
     /**
-     * Get the total lap time
-     * @return type
+     * Total time in seconds.
+     *
+     * @var float
      */
-    public function getTotalTime(){
+    protected $totalTime;
+
+    /**
+     * Totam distance in meters.
+     *
+     * @var float
+     */
+    protected $totalDistance;
+
+    /**
+     * Metres per second.
+     *
+     * @var float
+     */
+    protected $maxSpeed;
+
+    /**
+     * @var int
+     */
+    protected $totalCalories;
+
+    /**
+     * @var int
+     */
+    protected $avgHeartRate;
+
+    /**
+     * @var int
+     */
+    protected $maxHeartRate;
+
+    /**
+     * @var int
+     */
+    protected $cadence;
+
+    /**
+     * @var array<TrackPoint>
+     */
+    protected $trackPoints = [];
+
+    /**
+     * Get the total lap time.
+     *
+     * @return float
+     */
+    public function getTotalTime(): float
+    {
         return $this->totalTime;
     }
-    
+
     /**
-     * Get the total lap distance
-     * @return type
+     * Get the total lap distance.
+     *
+     * @return float
      */
-    public function getTotalDistance(){
+    public function getTotalDistance(): float
+    {
         return $this->totalDistance;
     }
-    
+
     /**
-     * Get the max speed achieved during the lap
-     * @return type
+     * Get the max speed achieved during the lap.
+     *
+     * @return float|null
      */
-    public function getMaxSpeed(){
+    public function getMaxSpeed(): ?float
+    {
         return $this->maxSpeed;
     }
-    
+
     /**
-     * Get the calories burnt during the lap
-     * @return type
+     * Get the calories burnt during the lap.
+     *
+     * @return float
      */
-    public function getTotalCalories(){
+    public function getTotalCalories(): ?float
+    {
         return $this->totalCalories;
     }
-    
+
     /**
-     * Get the array of track points
-     * @return type
+     * Get the array of track points.
+     *
+     * @return array
      */
-    public function getTrackPoints(){
+    public function getTrackPoints(): array
+    {
         return $this->trackPoints;
     }
-    
+
     /**
-     * Get a specific track point, by its number
-     * @param type $num
-     * @return type
+     * Get a specific track point, by its number.
+     *
+     * @param int $num
+     *
+     * @return TrackPoint
      */
-    public function getTrackPoint($num){
+    public function getTrackPoint(int $num): ?TrackPoint
+    {
         return (array_key_exists($num, $this->trackPoints)) ? $this->trackPoints[$num] : false;
     }
-    
+
     /**
-     * Set the total lap time (seconds)
-     * @param type $val
+     * Get last track point.
+     *
+     * @return TrackPoint
+     */
+    public function getLastTrackPoint(): ?TrackPoint
+    {
+        return $this->getTrackPoint(count($this->trackPoints) - 1);
+    }
+
+    /**
+     * Get the average heart rate achieved during the lap.
+     *
+     * @return int|null
+     */
+    public function getAvgHeartRate(): ?int
+    {
+        return $this->avgHeartRate;
+    }
+
+    /**
+     * Get the maximum heart rate achieved during the lap.
+     *
+     * @return int|null
+     */
+    public function getMaxHeartRate(): ?int
+    {
+        return $this->maxHeartRate;
+    }
+
+    /**
+     * Get the cadence achieved during the lap.
+     *
+     * @return int|null
+     */
+    public function getCadence(): ?int
+    {
+        return $this->cadence;
+    }
+
+    /**
+     * Set the total lap time (seconds).
+     *
+     * @param float $val
+     *
      * @return $this
      */
-    public function setTotalTime($val){
+    public function setTotalTime(float $val): Lap
+    {
         $this->totalTime = $val;
+
         return $this;
     }
-    
+
     /**
-     * Set the total lap distance (metres)
-     * @param type $val
+     * Set the total lap distance (metres).
+     *
+     * @param float $val
+     *
      * @return $this
      */
-    public function setTotalDistance($val){
+    public function setTotalDistance(float $val): Lap
+    {
         $this->totalDistance = $val;
+
         return $this;
     }
-    
+
     /**
-     * Set the max lap speed (metres per second)
-     * @param type $val
+     * Set the max lap speed (metres per second).
+     *
+     * @param float $val
+     *
      * @return $this
      */
-    public function setMaxSpeed($val){
+    public function setMaxSpeed(float $val): Lap
+    {
         $this->maxSpeed = $val;
+
         return $this;
     }
-    
+
     /**
-     * Set the total calories burnt
-     * @param type $val
+     * Set the total calories burnt.
+     *
+     * @param float $val
+     *
      * @return $this
      */
-    public function setTotalCalories($val){
+    public function setTotalCalories(float $val): Lap
+    {
         $this->totalCalories = $val;
+
         return $this;
     }
-    
+
     /**
-     * Add a track point to the lap
-     * @param \Waddle\TrackPoint $point
-     * @return \Waddle\TrackPoint
+     * Add a track point to the lap.
+     *
+     * @param TrackPoint $point
+     *
+     * @return TrackPoint
      */
-    public function addTrackPoint(TrackPoint $point){
-        $this->trackPoints[] = $point;     
+    public function addTrackPoint(TrackPoint $point): TrackPoint
+    {
+        $this->trackPoints[] = $point;
+
         return $point;
     }
-    
-    
+
+    /**
+     * Set the average heart rate in lap.
+     *
+     * @param int $val
+     *
+     * @return $this
+     */
+    public function setAvgHeartRate(int $val): Lap
+    {
+        $this->avgHeartRate = $val;
+
+        return $this;
+    }
+
+    /**
+     * Set the maximum heart rate in lap.
+     *
+     * @param int $val
+     *
+     * @return $this
+     */
+    public function setMaxHeartRate(int $val): Lap
+    {
+        $this->maxHeartRate = $val;
+
+        return $this;
+    }
+
+    /**
+     * Set the cadence in lap.
+     *
+     * @param type $val
+     *
+     * @return $this
+     */
+    public function setCadence($val): Lap
+    {
+        $this->cadence = $val;
+
+        return $this;
+    }
 }
