@@ -4,147 +4,252 @@ namespace Waddle;
 
 class TrackPoint
 {
-    
-    protected $time; // Timestamp of the time this point was recorded (generally every second)
-    protected $position = array(); // Array of lat/lon
-    protected $altitude; // Altitude in metres 
-    protected $distance; // Distance travelled so far in metres
-    
-    // Extensions that may or may not be present depending on device
-    protected $speed; // Metres per second
-    protected $heartRate; // Current hearthate
-    protected $calories; // Total calories burnt so far
-    
     /**
-     * Get the timestamp in a given format
-     * @param type $format
-     * @return type
+     * Timestamp of the time this point was recorded (generally every second).
+     *
+     * @var \DateTime
      */
-    public function getTime($format){
-        return ($this->time instanceof \DateTime) ? $this->time->format($format) : $this->time;
-    }
-    
-    /**
-     * Get either the lat/long array or a specific value from it, if "lat" or "long" is passed in
-     * @param type $type
-     * @return type
-     */
-    public function getPosition($type = null){
-        return ( !is_null($type) && array_key_exists($type, $this->position) ) ? $this->position[$type] : $this->position;
-    }
-    
-    /**
-     * Get the altitude
-     * @return type
-     */
-    public function getAltitude(){
-        return $this->altitude;
-    }
-    
-    /**
-     * Get the distance so far
-     * @return type
-     */
-    public function getDistance(){
-        return $this->distance;
-    }
-    
-    /**
-     * Get the current speed at this point
-     * @return type
-     */
-    public function getSpeed(){
-        return $this->speed;
-    }
-    
-    /**
-     * Get the current heart rate at this point
-     * @return type
-     */
-    public function getHeartRate(){
-        return $this->heartRate;
-    }
-    
-    
+    protected $time;
 
     /**
-     * Get the number of calories burnt so far
+     * @var array
+     */
+    protected $position = []; // Array of lat/lon
+
+    /**
+     * Altitude in metres.
+     *
+     * @var float
+     */
+    protected $altitude;
+
+    /**
+     * Distance travelled so far in metres.
+     *
+     * @var float
+     */
+    protected $distance;
+
+    /**
+     * Metres per second.
+     *
+     * @var float
+     */
+    protected $speed;
+
+    /**
+     * @var int
+     */
+    protected $heartRate;
+
+    /**
+     * @var int
+     */
+    protected $cadence;
+
+    /**
+     * @var int
+     */
+    protected $calories;
+
+    /**
+     * Get the timestamp in a given format.
+     *
+     * @param string $format
+     *
+     * @return \DateTime|null
+     */
+    public function getTime(string $format = null)
+    {
+        return ($this->time instanceof \DateTime && null !== $format)
+                ? $this->time->format($format)
+                : $this->time;
+    }
+
+    /**
+     * Get either the lat/long array or a specific value from it, if "lat" or "long" is passed in.
+     *
+     * @param type $type
+     *
      * @return type
      */
-    public function getCalories(){
+    public function getPosition($type = null)
+    {
+        return (!is_null($type) && array_key_exists($type, $this->position))
+                ? $this->position[$type]
+                : $this->position;
+    }
+
+    /**
+     * Get the altitude.
+     *
+     * @return float
+     */
+    public function getAltitude(): float
+    {
+        return $this->altitude;
+    }
+
+    /**
+     * Get the distance so far.
+     *
+     * @return float
+     */
+    public function getDistance(): float
+    {
+        return $this->distance;
+    }
+
+    /**
+     * Get the current speed at this point.
+     *
+     * @return float
+     */
+    public function getSpeed(): float
+    {
+        return $this->speed;
+    }
+
+    /**
+     * Get the current heart rate at this point.
+     *
+     * @return int
+     */
+    public function getHeartRate(): int
+    {
+        return $this->heartRate;
+    }
+
+    /**
+     * Get the current cadence rate at this point.
+     *
+     * @return int
+     */
+    public function getCadence(): int
+    {
+        return $this->cadence;
+    }
+
+    /**
+     * Get the number of calories burnt so far.
+     *
+     * @return int
+     */
+    public function getCalories(): int
+    {
         return $this->calories;
     }
 
     /**
-     * Set the timestamp of this point
+     * Set the timestamp of this point.
+     *
      * @param \DateTime $time
+     *
      * @return $this
      */
-    public function setTime(\DateTime $time){
-        $time->setTimezone( new \DateTimeZone( date_default_timezone_get() ) );
+    public function setTime(\DateTime $time): TrackPoint
+    {
+        $time->setTimezone(new \DateTimeZone(date_default_timezone_get()));
         $this->time = $time;
+
         return $this;
     }
-    
+
     /**
-     * Set the position array
+     * Set the position array.
+     *
      * @param array $val
+     *
      * @return $this
      */
-    public function setPosition(array $val){
+    public function setPosition(array $val): TrackPoint
+    {
         $this->position = $val;
+
         return $this;
     }
-    
+
     /**
-     * Set the altitude
-     * @param type $val
+     * Set the altitude.
+     *
+     * @param float $val
+     *
      * @return $this
      */
-    public function setAltitude($val){
+    public function setAltitude(float $val): TrackPoint
+    {
         $this->altitude = $val;
+
         return $this;
     }
-    
+
     /**
-     * Set the distance
-     * @param type $val
+     * Set the distance.
+     *
+     * @param float $val
+     *
      * @return $this
      */
-    public function setDistance($val){
+    public function setDistance($val): TrackPoint
+    {
         $this->distance = $val;
+
         return $this;
     }
-    
+
     /**
-     * Set the speed
-     * @param type $val
+     * Set the speed.
+     *
+     * @param float $val
+     *
      * @return $this
      */
-    public function setSpeed($val){
+    public function setSpeed(float $val): TrackPoint
+    {
         $this->speed = $val;
+
         return $this;
     }
-    
+
     /**
-     * Set the heart rate
-     * @param type $val
+     * Set the heart rate.
+     *
+     * @param int $val
+     *
      * @return $this
      */
-    public function setHeartRate($val){
+    public function setHeartRate(int $val): TrackPoint
+    {
         $this->heartRate = $val;
+
         return $this;
     }
-   
-    
+
     /**
-     * Set the calories burnt so far
-     * @param type $val
+     * Set the cadence rate.
+     *
+     * @param int $val
+     *
      * @return $this
      */
-    public function setCalories($val){
-        $this->calories = $val;
+    public function setCadence(int $val): TrackPoint
+    {
+        $this->cadence = $val;
+
         return $this;
     }
-    
+
+    /**
+     * Set the calories burnt so far.
+     *
+     * @param int $val
+     *
+     * @return $this
+     */
+    public function setCalories(int $val): TrackPoint
+    {
+        $this->calories = $val;
+
+        return $this;
+    }
 }
